Revert "Unpin polars"
This reverts commit 851d1f5814a72ed2808414540f4b18a2c80898b2.

+++ tests/ert/ui_tests/cli/test_field_parameter.py
@@ -94,8 +94,7 @@ def _compare_ensemble_params(
                 pl.col(c).map_elements(
                     lambda x: 0.0
                     if abs(x) < outlier_threshold
-                    else (abs(x) - outlier_threshold),
-                    return_dtype=pl.Float32,
+                    else (abs(x) - outlier_threshold)
                 )
             ).alias(c)
             for c in columns
@@ -106,10 +105,7 @@ def _compare_ensemble_params(
     outlier_percentage = truncated_df.select(
         [
             pl.concat_list(columns)
-            .map_elements(
-                lambda row: sum(1 for x in row if x != 0.0) / len(row),
-                return_dtype=pl.Float32,
-            )
+            .map_elements(lambda row: sum(1 for x in row if x != 0.0) / len(row))
             .alias("outlier_percentage")
         ]
     )["outlier_percentage"]
@@ -117,7 +113,7 @@ def _compare_ensemble_params(
     max_deviance = truncated_df.select(
         [
             pl.concat_list(columns)
-            .map_elements(lambda row: max(x for x in row), return_dtype=pl.Float32)
+            .map_elements(lambda row: max(x for x in row))
             .alias("max_deviance")
         ]
     )["max_deviance"]
